move id_map.py
